refactor: RankingPeriodItem 분리
