+++ frontend/src/components/RankingPeriodItem/type.ts
@@ -1,0 +1,13 @@
+import { RankingPeriod } from 'types/ranking';
+
+export type RankingPeriodItemProps = RankingPeriod & {
+  selected: boolean;
+  handleChooseRankingPeriod: () => void;
+};
+
+export type SelectListItemProps = Pick<RankingPeriodItemProps, 'selected'>;
+
+export type UseRankingPeriodItemProps = Pick<
+  RankingPeriodItemProps,
+  'startDate' | 'duration'
+>;
