+++ frontend/src/components/RankingPeriodsList/index.tsx
@@ -1,18 +1,15 @@
-import { RankingPeriodItemProps, SelectListItemProps, SelectListProps } from './type';
+import { SelectListProps } from './type';
 import { useGetRankingPeriods } from 'apis/rankingApi';
 import { useState } from 'react';
 import { BsFillCaretDownFill } from 'react-icons/bs';
 import styled, { css } from 'styled-components';
-import { addDays, parseTime } from 'utils';
+import { addDays, dateYMDFormatParsing } from 'utils';
 
 import { FlexBox } from 'components/@shared/FlexBox';
 
-import { RANKING_DURATION } from 'constants/domain';
+import { RankingPeriodItem } from 'components/RankingPeriodItem';
 
-const dateFormatParsing = (dateString: string) => {
-  const { year, month, date } = parseTime(dateString);
-  return `${year}.${month}.${date}`;
-};
+import { RANKING_DURATION } from 'constants/domain';
 
 export const RankingPeriodsList = () => {
   const [selectedPeriodIndex, setSelectedPeriodIndex] = useState(0);
@@ -34,8 +31,8 @@ export const RankingPeriodsList = () => {
 
   const { startDate, duration } = rankingPeriodsData.data[selectedPeriodIndex];
 
-  const startDateString = dateFormatParsing(startDate);
-  const endDateString = dateFormatParsing(
+  const startDateString = dateYMDFormatParsing(startDate);
+  const endDateString = dateYMDFormatParsing(
     String(addDays(new Date(Date.parse(startDate)), RANKING_DURATION[duration])),
   );
 
@@ -59,27 +56,6 @@ export const RankingPeriodsList = () => {
   );
 };
 
-const RankingPeriodItem = ({
-  rankingPeriodId,
-  startDate,
-  duration,
-  selected,
-  handleChooseRankingPeriod,
-}: RankingPeriodItemProps) => {
-  const startDateString = dateFormatParsing(startDate);
-  const endDateString = dateFormatParsing(
-    String(addDays(new Date(Date.parse(startDate)), RANKING_DURATION[duration])),
-  );
-
-  return (
-    <SelectListItem selected={selected} onClick={handleChooseRankingPeriod}>
-      <button>
-        {startDateString} ~ {endDateString}
-      </button>
-    </SelectListItem>
-  );
-};
-
 const SelectPeriod = styled(FlexBox)`
   ${({ theme }) => css`
     width: fit-content;
@@ -88,7 +64,6 @@ const SelectPeriod = styled(FlexBox)`
     font-weight: bold;
     font-size: 1.2rem;
     text-align: center;
-    /* border-radius: 10px; */
     border-radius: 20px;
     background-color: ${theme.surface};
     color: ${theme.primary};
@@ -104,24 +79,5 @@ const SelectList = styled.ul`
     overflow-y: scroll;
     position: relative;
     right: -0.5rem;
-  `}
-`;
-
-const SelectListItem = styled.li<SelectListItemProps>`
-  ${({ theme, selected }) => css`
-    width: fit-content;
-    height: 2.5rem;
-    line-height: 2.5rem;
-    text-align: center;
-    background-color: ${theme.surface};
-    border: 1px solid ${theme.border};
-    button {
-      font-size: 1.2rem;
-      ${
-        selected &&
-        css`
-          color: ${theme.primary};
-        `
-      }
   `}
 `;
