+++ frontend/src/components/RankingPeriodsList/type.ts
@@ -1,12 +1,3 @@
-import { RankingPeriod } from 'types/ranking';
-
-export type RankingPeriodItemProps = RankingPeriod & {
-  selected: boolean;
-  handleChooseRankingPeriod: () => void;
-};
-
 export type SelectListProps = {
   show: boolean;
 };
-
-export type SelectListItemProps = Pick<RankingPeriodItemProps, 'selected'>;
