+++ frontend/src/components/index.ts
@@ -54,4 +54,5 @@ export * from 'components/ErrorFallbackLoginButton';
 export * from 'components/ErrorFallbackSubscriptionButton';
 export * from 'components/ErrorFallbackNotificationMessage';
 export * from 'components/ScrollToTop';
+export * from 'components/RankingPeriodItem';
 export * from 'components/RankingPeriodsList';
